\docType{data}
\name{air_passengers}
\alias{air_passengers}
\title{Number of air passengers}
\format{A monthly time series, in thousands, with column \code{year} indicating year and the names \code{Jan:Dec} the month.
    }
\usage{
  data(air_passengers)
}
\description{
  A slightly modified version of the classic Box and Jenkins data set of monthly totals of international airline passengers, 1949 to 1960.
}
\references{
Box, G. E. P., Jenkins, G. M. and Reinsel, G. C. (1976) Time Series Analysis, Forecasting and Control. Third Edition. Holden-Day. Series G.
}
\seealso{
\link[datasets]{AirPassengers}
}
