\docType{data}
\name{dat}
\alias{dat}
\title{dat: Four random walks}
\format{Four random walks of length 100, described by:
\describe{
\item{time}{The time points (1--100)}
\item{variable}{The identifier for each random walk (A, B, C, D)}
\item{value}{The value of the time series}
}}
\usage{
  data(dat)
}
\description{
  Four random walks, generated with independent N(0, 1) errors. Used to illustrate ggplot2 vs base graphics.
}

