% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/galaxies.R
\docType{data}
\name{galaxies}
\alias{galaxies}
\title{Velocities for 81 Galaxies}
\format{
An object of class \code{numeric} of length 82.
}
\usage{
data(galaxies)
}
\description{
A numeric vector of transformed velocities in km/sec divided by 1000 of 82 galaxies from 6
well-separated conic sections of an unfilled survey of the
Corona Borealis region.
}
\references{
Roeder, K. (1990) Density estimation with
confidence sets exemplified by superclusters and voids in galaxies.
Journal of the American Statistical Association 85, 617–624.

Postman, M., Huchra, J. P. and Geller, M. J. (1986) Probes of large-scale
structures in the Corona Borealis region. Astronomical Journal 92, 1238–1247.
}
\keyword{datasets}
