\docType{data}
\name{senate}
\alias{senate}
\title{Senate data}
\format{A data frame with 107 rows and 4 variables. The variables are:
    \describe{
      \item{year}{Year}
      \item{election_result}{The margin in the election}
      \item{presidential_approval}{The presidential approval rating}
      \item{poll_average}{The average of polls}
    }}
\usage{
  data(senate)
}
\source{
\url{https://github.com/fivethirtyeight/data/tree/master/early-senate-polls}
}
\description{
  This data is used for the figures in this article: \url{http://fivethirtyeight.com/features/early-senate-polls-have-plenty-to-tell-us-about-november/}.
}

