\docType{data}
\name{titanic}
\alias{titanic}
\title{Titanic data}
\format{A data frame with 891 rows and 12 variables. The variables are:
    \describe{
      \item{PassengerId}{The ID of the passenger}
      \item{Survived}{1 indicates survival}
      \item{Pclass}{The class in which the passenger travelled}
      \item{Name}{The name }
      \item{Sex}{The sex}
      \item{Age}{The age}
      \item{SibSp}{Number of siblings/spouses aboard}
      \item{Parch}{Number of parents/children aboard}
      \item{Ticket}{Ticket number}
      \item{Fare}{Passenger fare}
      \item{Cabin}{Cabin}
      \item{Embarked}{Port of embarkation}
    }}
\usage{
  data(titanic)
}
\description{
  This dataset is a transformed version of what is available in \code{\link[titanic]{titanic_train}}.
}

