% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/galaxies.R
\docType{data}
\name{galaxies}
\alias{galaxies}
\title{Velocities for 81 Galaxies}
\format{
An object of class \code{numeric} of length 81.
}
\usage{
data(galaxies)
}
\description{
A numeric vector of velocities in km/sec of 81 galaxies divided by 1000 from 6
well-separated conic sections of an unfilled survey of the
Corona Borealis region. Multimodality in such surveys is evidence
for voids and superclusters in the far universe.
}
\references{
Roeder, K. (1990) Density estimation with
confidence sets exemplified by superclusters and voids in galaxies.
Journal of the American Statistical Association 85, 617–624.

Postman, M., Huchra, J. P. and Geller, M. J. (1986) Probes of large-scale
structures in the Corona Borealis region. Astronomical Journal 92, 1238–1247.
}
\keyword{datasets}
